# Refactor batch_hash for clarity.